Move UI to model
